pyproject.toml
[project]
name = "codefusion-assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dj-database-url>=2.3.0",
    "django>=5.2",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
